Move DocumentationVersion to own file
